backend/requirements/local.txt
-r ./base.txt

coverage==6.5.0
safety==2.3.1
black==22.10.0
flake8-black==0.3.3

django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework-stubs==1.7.0
factory-boy==3.2.1
Faker==15.1.1
flake8==5.0.4
flake8-isort==5.0.0

bandit==1.7.4

pytest==7.1.0
pytest-django==4.5.2
pytest-sugar==0.9.5

pycodestyle==2.9.1
pyflakes==2.5.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7

types-six==1.16.21
